tests: updated for the new behavior of handshake
Previously at handshake we would negotiate a ciphersuite and certificate
and later figure out a signature algorithm. Now we negotiate all at once,
so we no longer reach situations where mid-way of handshake we figure we
have no signature algorithm to use. Update the test cases relying on that
behavior to account the new one.

Signed-off-by: Nikos Mavrogiannopoulos <nmav@redhat.com>